refactor(接口测试): 优化批量执行失败停止报告处理
